requirements.txt
git+https://github.com/aleximmer/Laplace.git@main#egg=laplace-torch
git+https://github.com/facebookresearch/sam2.git@main#egg=SAM-2
git+https://github.com/facebookresearch/segment-anything.git@main#egg=segment_anything

transformers
datasets
pandas
numpy 
torch
matplotlib 
pillow
opencv-python
pycocotools
nibabel
python-dotenv
pyaml
tabulate
wget

scikit-learn
scipy
tqdm
torchvision
onnx
pytorch-lightning
importlib-metadata>=6.8
tensordict